+++ tsp/algorithms/refinement/2opt/2opt.c
@@ -88,7 +88,7 @@ double getOpt2OptMove(const TSPInstance* inst, const TSPSolution* sol, int* opti
 * OR int execution seconds
 * NB: this method will perform the best move, not just the first one that improve the solution cost
 */
-int opt2(const Settings* set, const TSPInstance* inst, TSPSolution* sol){
+double opt2(const Settings* set, const TSPInstance* inst, TSPSolution* sol){
 
 	clock_t start = clock();
 	double ls = -1;
