+++ tsp/algorithms/refinement/2opt/2opt.h
@@ -7,7 +7,7 @@
 
 #include "../../../tsp.h"
 
-int opt2(const Settings*, const TSPInstance*, TSPSolution*);
+double opt2(const Settings*, const TSPInstance*, TSPSolution*);
 
 void opt2move(int, int, const TSPInstance*, TSPSolution*);
